refactor: sync local edits

+++ configs/.config/eww/scripts/countdowner.sh
@@ -3,6 +3,11 @@
 PIDFILE="/tmp/eww_countdowner.pid"
 
 countdown() {
+  if [[ -f "$PIDFILE" ]]; then
+    echo "There is another countdowner running"
+    exit 1
+  fi
+
   echo "$BASHPID" > "$PIDFILE"
   trap pause_now SIGUSR1
   trap resume_now SIGUSR2
@@ -10,13 +15,16 @@ countdown() {
   PID=$(cat "$PIDFILE")
   echo "$PID"
 
-  local total_min=$(echo "$1 * 60 - 1" | bc | cut -d'.' -f1)
-  local min=59
-  local hour=$(($total_min / 60))
+  local hour=$1
+  local min=$2
 
   eww open countdowner
 
-  while [ $min -ge 0 ]; do
+  while [[ "$hour" -ge 0 ]]; do
+    if [[ "$min" -eq 0 ]]; then
+      ((hour--))
+      min=59
+    fi
     eww update countdowner_time="$hour:$min"
     ((min--))
     sleep 1m
@@ -29,8 +37,7 @@ quit() {
     echo "$PID"
     if kill -0 "$PID" 2>/dev/null; then
       kill -TERM "$PID"
-    else
-      exit 1
+
     fi
     rm -f "$PIDFILE"
     eww close countdowner
@@ -40,12 +47,12 @@ quit() {
 }
 
 print_help(){
-printf "SYNOPSIS:\n
-    ./countdowner.sh [OPTION [...]]\n
-OPTIONS:\n\n"
-printf "  %-30s %s\n" "-s TIME(hour), --start TIME(hour)" "- Set TIME(hour) to countdown"
-printf "  %-30s %s\n" "-q, --quit" "- Quit the countdown"
-printf "  %-30s %s\n" "-h, --help" "- Show this message again"
+  printf "SYNOPSIS:\n
+      ./countdowner.sh [OPTION [...]]\n
+  OPTIONS:\n\n"
+  printf "  %-30s %s\n" "-s hour min, --start hour min" "- Set TIME(hour) to countdown"
+  printf "  %-30s %s\n" "-q, --quit" "- Quit the countdown"
+  printf "  %-30s %s\n" "-h, --help" "- Show this message again"
 }
 
 OPTIONS=$(getopt --options=s:,q,h --longoptions=start:,quit,help --name "$0" -- "$@")
@@ -55,7 +62,17 @@ eval set -- "$OPTIONS"
 while true; do
 case "$1" in
   -s|--start)
-    countdown "$2" &
+    time_arg=$2
+
+    if [[ $time_arg == *:* ]]; then
+      hour=${time_arg%%:*}
+      min=${time_arg##*:}
+    else 
+      hour=$time_arg
+      min=0
+    fi
+
+    countdown "$hour" "$min" &
     shift 2
   ;;
 -q|--quit)
